# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI-powered resume engineering using Hexagonal Architecture. Tailors resumes to specific job descriptions using LLM technology.",
        "title": "Chameleon Vitae API",
        "contact": {
            "name": "Chameleon Vitae Team",
            "url": "https://github.com/SeltikHD/chameleon-vitae"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/resumes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of resumes for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resumes"
                ],
                "summary": "List resumes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ListResumesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new resume draft from a job description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resumes"
                ],
                "summary": "Create resume",
                "parameters": [
                    {
                        "description": "Resume data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.CreateResumeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ResumeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/resumes/{resumeID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a specific resume with full content",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resumes"
                ],
                "summary": "Get resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "resumeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ResumeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resume not found",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a resume",
                "tags": [
                    "resumes"
                ],
                "summary": "Delete resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "resumeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resume not found",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/resumes/{resumeID}/content": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the status or content of a resume for manual adjustments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resumes"
                ],
                "summary": "Update resume status/content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "resumeID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resume update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.UpdateResumeContentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ResumeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resume not found",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/resumes/{resumeID}/pdf": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates and downloads a PDF file of the resume",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "resumes"
                ],
                "summary": "Generate PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "resumeID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "modern",
                        "description": "Template name",
                        "name": "template",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Force regeneration ignoring cache",
                        "name": "force_regenerate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resume not found",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Resume not ready for PDF",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_primary_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_adapters_primary_http.CreateResumeRequest": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string",
                    "example": "Awesome Corp"
                },
                "job_description": {
                    "type": "string",
                    "example": "## Senior Backend Engineer..."
                },
                "job_title": {
                    "type": "string",
                    "example": "Senior Backend Engineer"
                },
                "job_url": {
                    "type": "string",
                    "example": "https://linkedin.com/jobs/12345"
                },
                "target_language": {
                    "type": "string",
                    "example": "en"
                }
            }
        },
...
        "internal_adapters_primary_http.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_adapters_primary_http.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "internal_adapters_primary_http.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid email format"
                }
            }
        },
        "internal_adapters_primary_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_adapters_primary_http.ErrorBody"
                }
            }
        },
...
        "internal_adapters_primary_http.ListResumesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_adapters_primary_http.ResumeResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "internal_adapters_primary_http.UpdateResumeContentRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string",
                    "example": "Made adjustments to summary"
                },
                "status": {
                    "type": "string",
                    "example": "reviewed"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Firebase JWT token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}